+++ R/airsisDump_createMonitorObject.R
@@ -46,7 +46,7 @@ airsisDump_createMonitorObject <- function(filepath,
   # Loop over monitor tibble list (mostly verbatim from wrcc_createMonitorObject)
   for ( name in names(tblList) ) {
     
-    logger.info("Processing data for %s ...", name)
+    logger.debug("Processing data for %s ...", name)
     
     tbl <- tblList[[name]]