+++ R/airsis_createMonitorObject.R
@@ -51,6 +51,7 @@
 #' @seealso \code{\link{airsis_createDataDataframe}}
 #' @examples
 #' \dontrun{
+#' initializeMazamaSpatialUtils()
 #' usfs_1013 <- airsis_createMonitorObject(20150301, 20150831, 'USFS', unitID='1013')
 #' monitorLeaflet(usfs_1013)
 #' }